Move ingest job status updates to core module
