+++ java/core/src/main/java/sleeper/core/tracker/ingest/job/query/IngestJobAcceptedStatus.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.ingest.core.job.status;
+package sleeper.core.tracker.ingest.job.query;
 
 import java.time.Instant;
 import java.util.Objects;
