+++ java/core/src/main/java/sleeper/core/tracker/ingest/job/query/IngestJobAddedFilesStatus.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.ingest.core.job.status;
+package sleeper.core.tracker.ingest.job.query;
 
 import sleeper.core.record.process.status.ProcessStatusUpdate;
