+++ java/core/src/main/java/sleeper/core/tracker/ingest/job/query/IngestJobValidatedStatus.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.ingest.core.job.status;
+package sleeper.core.tracker.ingest.job.query;
 
 /**
  * An ingest job info status that has validation information about the job.
